Randomize Fishing and Headbutt Encounters
This randomizes the Fishing and Headbutt encounters when the randomize wild encounters setting is toggled. Fishing encounters do not respect time of day and are simply 11 random encounters across the three rods in each of the 12 defined fishing tables (the 10 areas plus the two swarm tables is I think what's present). Headbutt encounters are simply 13 random Pokemon put in the original slots so finding rare trees will still work as it did in vanilla. Neither of these impact encounter probability.

This push also removes the move names from the nice item names for the TM items. This should make the spoiler log cleaner and should make !hint and !getitem more intuitive to use. The logic rules were correctly updated to prevent this from breaking compatibility with access to the Rock Smash TM.